# Conflicts:
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/.vs/Projet_Huchon_Salemi_3I/v16/.suo
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/DAO/VehiculeDAO.cs
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/MainWindow.xaml.cs
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/bin/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.dll
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/bin/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.dll.locked
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/bin/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.pdb
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.csproj.AssemblyReference.cache
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.csproj.CoreCompileInputs.cache
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.dll
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.pdb
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I_MarkupCompile.cache

+++ Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/metier/Balade.cs
@@ -80,7 +80,7 @@ namespace Projet_Huchon_Salemi_3I.metier
 
             foreach (Vehicule i in balade.listeVehicule)
             {
-                totalNbrePlaces = totalNbrePlaces + i.NbrePlacesMembre;
+                totalNbrePlaces = totalNbrePlaces + (int) i.NbrePlacesMembre;
             }
             return totalNbrePlaces;
         }
@@ -118,7 +118,7 @@ namespace Projet_Huchon_Salemi_3I.metier
 
             foreach(Vehicule i in balade.listeVehicule)
             {
-                totalNbrePlaces = totalNbrePlaces + i.NbrePlacesVelo;
+                totalNbrePlaces = totalNbrePlaces + (int) i.NbrePlacesVelo;
             }
             return totalNbrePlaces;
         }
